server/requirements.txt
aiostream==0.6.4
annotated-types==0.7.0
anyio==4.6.2.post1
beanie==1.27.0
certifi==2024.8.30
cffi==1.17.1
charset-normalizer==3.3.2
click==8.1.7
colorama==0.4.6
contextlib2==21.6.0
cryptography==43.0.1
dnspython==2.6.1
email_validator==2.2.0
fastapi==0.115.5
fastapi-cli==0.0.5
fire==0.7.0
gunicorn==23.0.0
h11==0.14.0
httpcore==1.0.6
httptools==0.6.4
httpx==0.27.2
idna==3.10
inflection==0.5.1
injector==0.22.0
itsdangerous==2.2.0
Jinja2==3.1.4
joblib==1.4.2
lazy-model==0.2.0
markdown-it-py==3.0.0
MarkupSafe==3.0.2
mdurl==0.1.2
motor==3.6.0
numpy==1.26.4
oauthlib==3.2.2
orjson==3.10.11
packaging==24.1
psycopg==3.2.2
psycopg-binary==3.2.2
pycparser==2.22
pydantic==2.9.2
pydantic-extra-types==2.10.0
pydantic-settings==2.6.1
pydantic_core==2.23.4
Pygments==2.18.0
PyJWT==2.9.0
pymongo==4.9.2
python-decouple==3.8
python-dotenv==1.0.1
python-http-client==3.3.7
python-multipart==0.0.17
pytz==2024.2
PyYAML==6.0.2
rav==0.0.9
requests==2.32.3
requests-oauthlib==2.0.0
rich==13.9.2
scikit-learn==1.4.2
scipy==1.13.0
sendgrid==6.11.0
shellingham==1.5.4
slippers==0.6.2
sniffio==1.3.1
sqlparse==0.5.1
starkbank-ecdsa==2.2.0
starlette==0.41.2
stripe==10.12.0
termcolor==2.4.0
threadpoolctl==3.5.0
toml==0.10.2
typeguard==2.13.3
typer==0.13.0
typing_extensions==4.12.2
tzdata==2024.1
ujson==5.10.0
uritemplate==4.1.1
urllib3==2.2.3
uvicorn==0.32.0
watchfiles==0.24.0
websockets==14.1
whitenoise==6.7.0
passlib
jose
